# internal/appview/openapi.yaml
openapi: 3.0.3
info:
  title: HashPost AppView API
  description: HashPost AppView provides a unified API for accessing HashPost data aggregated from the PDS
  version: 1.0.0
  contact:
    name: HashPost Team
    email: team@hashpost.com
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://appview.hashpost.com
    description: Production server
paths:
...
  /api/v1/posts/{id}/vote:
    post:
      summary: Vote on post
      description: Vote up or down on a post
      operationId: voteOnPost
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vote_type
              properties:
                vote_type:
                  type: string
                  enum: [up, down]
                  description: Type of vote
      responses:
        '200':
          description: Vote successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    delete:
      summary: Remove vote from post
      description: Remove user's vote from a post
      operationId: removeVoteFromPost
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID
      responses:
        '200':
          description: Vote removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      summary: Get user's vote on post
      description: Get the authenticated user's vote on a post
      operationId: getUserVoteOnPost
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID
      responses:
        '200':
          description: User's vote status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/comments/{id}/vote:
    post:
      summary: Vote on comment
      description: Vote up or down on a comment
      operationId: voteOnComment
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Comment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vote_type
              properties:
                vote_type:
                  type: string
                  enum: [up, down]
                  description: Type of vote
      responses:
        '200':
          description: Vote successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    delete:
      summary: Remove vote from comment
      description: Remove user's vote from a comment
      operationId: removeVoteFromComment
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Comment ID
      responses:
        '200':
          description: Vote removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      summary: Get user's vote on comment
      description: Get the authenticated user's vote on a comment
      operationId: getUserVoteOnComment
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Comment ID
      responses:
        '200':
          description: User's vote status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # Computed data endpoints
...
components:
  schemas:
...
    VoteResponse:
      type: object
      required:
        - vote_type
        - upvotes
        - downvotes
        - score
      properties:
        vote_type:
          type: string
          enum: [up, down, ""]
          description: Type of vote (empty string if no vote)
        upvotes:
          type: integer
          minimum: 0
          description: Number of upvotes
        downvotes:
          type: integer
          minimum: 0
          description: Number of downvotes
        score:
          type: integer
          description: Net score (upvotes - downvotes)

    # Subscription schemas
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from PDS authentication
...